+++ CHANGELOG.md
@@ -8,6 +8,13 @@ as well as to [Module version numbering](https://go.dev/doc/modules/version-numb
 
 ## [Unreleased](https://github.com/goyek/goyek/compare/v0.1.1...HEAD)
 
+### Added
+
+- Add `color.Disable` disables colorizing the output.
+- Add `color.Force` forces colorizing the output.
+- Add `-no-color` and `-force-color` flags in `boot.Main`
+  to give more control on colorizing the output.
+
 ## [0.1.1](https://github.com/goyek/goyek/releases/tag/v0.1.1) - 2022-11-06
 
 This release bumps `goyek` to `2.0.0-rc.9`.